+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v3.1.3-1.18.2] - 2022-03-18
+### Added
+- Added a description to magnum torch items
+### Changed
+- Tweaked spawn types config settings to better explain what certain types actually do
+### Fixed
+- Fixed wandering traders and zombified piglins from portals not properly being prevented from spawning
+
 ## [v3.1.2-1.18.2] - 2022-03-15
 ### Changed
 - Black-/whitelist now appear on the inventory tooltip when used